完善项目架构

+++ src/main/java/cn/zcn/distributed/lock/redis/RedisSubscriptionListener.java
@@ -3,7 +3,7 @@ package cn.zcn.distributed.lock.redis;
 public interface RedisSubscriptionListener {
     void onMessage(byte[] channel, byte[] message);
 
-    void onSubscribe(byte[] channel, int subscribedChannels);
+    void onSubscribe(byte[] channel, long subscribedChannels);
 
-    void onUnsubscribe(byte[] channel, int subscribedChannels);
+    void onUnsubscribe(byte[] channel, long subscribedChannels);
 }
